added Menu Item From Add Screen functionality

+++ src/test/java/testCases/SalesModule/BeverageServiceTest.java
@@ -35,11 +35,10 @@ public class BeverageServiceTest extends BaseClass {
 	@Test(priority = 1, groups = { "Regression", "All" })
 	public void beveageservice() {
 		String eventNo = servicesNewApi.newServiceEventId(loginId, service);
-		servicesFlow.openServiceRequestFromEventListing(eventNo, service);
 
-		Assert.assertTrue(service.stream().anyMatch(s -> servicesPage.getServiceHdr().contains(s)),
-				"BeverageService not Mapped/Service not present in the Service list.");
-
+		if (!servicesFlow.openServiceRequestFromEventListing(eventNo, service)) {
+			return;
+		}
 		beverageServicePage.addBeverageItems();
 
 		beverageServicePage.validateItems();
@@ -51,11 +50,10 @@ public class BeverageServiceTest extends BaseClass {
 	@Test(priority = 2, groups = { "Regression", "All" })
 	public void editBeverageItem() {
 		String eventNo = servicesNewApi.newServiceEventId(loginId, service);
-		servicesFlow.openServiceRequestFromEventListing(eventNo, service);
 
-		Assert.assertTrue(service.stream().anyMatch(s -> servicesPage.getServiceHdr().contains(s)),
-				"BeverageService not Mapped/Service not present in the Service list.");
-
+		if (!servicesFlow.openServiceRequestFromEventListing(eventNo, service)) {
+			return;
+		}
 		beverageServicePage.addBeverageItems();
 		beverageServicePage.editQuantity();
 		servicesPage.clickServiceSave();
@@ -64,17 +62,31 @@ public class BeverageServiceTest extends BaseClass {
 	@Test(priority = 3, groups = { "Regression", "All" })
 	public void deleteBeverageItem() {
 		String eventNo = servicesNewApi.newServiceEventId(loginId, service);
-		servicesFlow.openServiceRequestFromEventListing(eventNo, service);
 
-		Assert.assertTrue(service.stream().anyMatch(s -> servicesPage.getServiceHdr().contains(s)),
-				"BeverageService not Mapped/Service not present in the Service list.");
-
+		if (!servicesFlow.openServiceRequestFromEventListing(eventNo, service)) {
+			return;
+		}
 		beverageServicePage.addBeverageItems();
 		String itemName = beverageServicePage.getBeverageItemName();
 		beverageServicePage.deleteItem();
 		servicesPage.clickAlertOk();
 
 		Assert.assertFalse(servicesPage.isItemPresent(itemName), "Item is Not deleted");
+	}
+
+	@Test(priority = 4, groups = { "Regression", "All" })
+	public void verifyBeverageItemUsingAdd() {
+		String eventNo = servicesNewApi.newServiceEventId(loginId, service);
+
+		if (!servicesFlow.openServiceRequestFromEventListing(eventNo, service)) {
+			return;
+		}
+		servicesPage.clickAddbtn();
+		beverageServicePage.clickExpandAll();
+		beverageServicePage.addItemsFromAdd("10");
+		servicesPage.clickAddSave();
+		servicesPage.clickAddClose();
+
 	}
 
 	@AfterMethod(alwaysRun = true)
